+++ packages/core/src/index.ts
@@ -5,7 +5,7 @@ export * from "./types";
 export * from "./utils";
 
 // Constants
-export { ComponentTypes, FormConstants, FIELD_PARENT_PREFIX } from "./constants";
+export { ComponentTypes, FormConstants } from "./constants";
 
 // Strings
 export { FormStrings } from "./strings";
@@ -52,10 +52,9 @@ export {
   InitOnEditFormState,
   ShowField,
   GetFieldsToRender,
-  SortOptions,
-} from "./helpers/InlineFormHelper";
+} from "./helpers/FormosaicHelper";
 
-export { SortOptions as SortDropdownOptions } from "./helpers/FieldHelper";
+export { SortOptions, SortOptions as SortDropdownOptions } from "./helpers/FieldHelper";
 
 export {
   detectDependencyCycles,
@@ -112,23 +111,12 @@ export { FieldArray } from "./components/FieldArray";
 export type { IFieldArrayProps } from "./components/FieldArray";
 
 // Components
-export { Formosaic } from "./components/InlineForm";
-export { FormFields } from "./components/InlineFormFields";
+export { Formosaic } from "./components/Formosaic";
+export { FormFields } from "./components/FormosaicFields";
 export { FieldWrapper } from "./components/FieldWrapper";
 export { default as RenderField } from "./components/RenderField";
 export { default as ConfirmInputsModal } from "./components/ConfirmInputsModal";
 export { FormErrorBoundary } from "./components/FormErrorBoundary";
-
-// DevTools
-export { FormDevTools } from "./components/FormDevTools";
-export type { IFormDevToolsProps } from "./components/FormDevTools";
-
-// RJSF Schema Import/Export
-export { fromRjsfSchema, toRjsfSchema } from "./utils/rjsf";
-export type { IJsonSchemaNode, IRjsfUiSchema, IRjsfConvertOptions } from "./utils/rjsf";
-
-// Zod Schema Import
-export { zodSchemaToFieldConfig } from "./utils/zodSchemaImport";
 
 // Lazy Field Registry
 export { createLazyFieldRegistry } from "./utils/lazyFieldRegistry";
@@ -161,32 +149,3 @@ export {
 
 // Expression Engine
 export { evaluateExpression, extractExpressionDependencies, extractFunctionDependencies } from "./helpers/ExpressionEngine";
-
-// Rule Tracing / Debugging
-export {
-  enableRuleTracing,
-  disableRuleTracing,
-  traceRuleEvent,
-  getRuleTraceLog,
-  clearRuleTraceLog,
-  isRuleTracingEnabled,
-} from "./helpers/RuleTracer";
-export type { IRuleTraceEvent } from "./helpers/RuleTracer";
-
-// Render Tracking
-export {
-  trackRender,
-  flushRenderCycle,
-  getRenderCounts,
-  getLastRenderedFields,
-  getTotalFormRenders,
-  resetRenderTracker,
-} from "./helpers/RenderTracker";
-
-// Event Timeline
-export {
-  logEvent,
-  getTimeline,
-  clearTimeline,
-} from "./helpers/EventTimeline";
-export type { ITimelineEvent, TimelineEventType } from "./helpers/EventTimeline";